BREAKING CHANGE: Rewrite generator script
The old design was bloated and relied on poor practices.

This version features improved scripting techniques and a cleaner syntax
that makes it easier to read and maintain.

The breaking change concerns the removal of features that fall outside
the scope of this project, such as previewing the colour scheme in the
running terminal.

The new generator does only one thing: prints to stdout a ready-to-use
theme, by parsing scheme data through a template.