vim: Add g:tempus_enforce_background_color option
Add a `g:tempus_enforce_background_color` option to set the background
color for vim used in a terminal.

The default remains to allow the theme to play well with transparent
terminals.

To force the background to be set, this option can be added to `.vimrc`:

```vim
let g:tempus_enforce_background_color=1
```

This allows for post colorscheme tweaks in the `.vimrc` using the
background color, like:

```vim
" Make the "~" characters after the end of buffer disappear
hi EndOfBuffer guifg=bg ctermfg=bg
```