EXPERIMENTAL: Add simplified Modus themes
These schemes are only meant to offer a facility to those who wish to
run a simplified version of the Modus themes for Emacs in their terminal
in Vim, etc.[0][1][2]

I have no plan to merge this branch into trunk: the Tempus themes and
the Modus themes will remain separate projects: they are designed
differently and their scope is not the same.

Furthermore, this branch will be deleted when/if I develop a fully
fledged exporter/generator for the Modus themes.

[0]: <https://gitlab.com/protesilaos/modus-themes>.
[1]: <https://gitlab.com/protesilaos/dotfiles/-/issues/31>.
[2]: <https://gitlab.com/protesilaos/modus-themes/-/issues/116>.